Create music module
