+++ json_files/albums.json
@@ -1,1 +1,1 @@
-[{"id":99,"name":"unreal unearth","publish_date":"18 August 2023","cover_state":"good","on_spotify":true,"archived":false},{"id":530,"name":"haha","publish_date":"aa","cover_state":"a","on_spotify":false,"archived":false},{"id":332,"name":"haha","publish_date":"12","cover_state":"bad","on_spotify":true,"archived":true},{"id":349,"name":"baba","publish_date":"12","cover_state":"bad","on_spotify":false,"archived":false},{"id":24,"name":"ivy","publish_date":"12","cover_state":"good","on_spotify":true,"archived":true},{"id":574,"name":"nope","publish_date":"21","cover_state":"good","on_spotify":true,"archived":true}]
+[{"id":365,"name":"unreal unearth","publish_date":"18 August 2023","cover_state":"good","on_spotify":true,"archived":false},{"id":747,"name":"haha","publish_date":"aa","cover_state":"a","on_spotify":false,"archived":false},{"id":346,"name":"haha","publish_date":"12","cover_state":"bad","on_spotify":true,"archived":true},{"id":524,"name":"baba","publish_date":"12","cover_state":"bad","on_spotify":false,"archived":false},{"id":556,"name":"ivy","publish_date":"12","cover_state":"good","on_spotify":true,"archived":true},{"id":295,"name":"nope","publish_date":"21","cover_state":"good","on_spotify":true,"archived":true},{"id":687,"name":"I live in fear","publish_date":"10","cover_state":"bad","on_spotify":true,"archived":true},{"id":56,"name":"nati matek","publish_date":"12","cover_state":"good","on_spotify":true,"archived":false}]
