+++ json_files/genres.json
@@ -1,1 +1,1 @@
-[{"id":260,"name":"alt rock"},{"id":507,"name":"a"},{"id":84,"name":"gaga"},{"id":464,"name":"hitler"},{"id":470,"name":"siblings"},{"id":328,"name":"pop"}]
+[{"id":913,"name":"alt rock"},{"id":840,"name":"a"},{"id":148,"name":"gaga"},{"id":451,"name":"hitler"},{"id":931,"name":"siblings"},{"id":328,"name":"pop"},{"id":279,"name":"depressed"},{"id":41,"name":"luo"}]
